+++ chopin-typescript/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.1.1] chopin-typescript
 ### Releases
 - [NPM v0.1.1] https://www.npmjs.com/package/@chopinframework/sdk/v/0.1.1 - chopin-typescript
+
+## 2025-03-07 01:37:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.513.2 (2.545.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.2] chopin-typescript
+### Releases
+- [NPM v0.1.2] https://www.npmjs.com/package/@chopinframework/sdk/v/0.1.2 - chopin-typescript